+++ alpha/apps/kaltura/lib/webservices/APIErrors.class.php
@@ -424,9 +424,11 @@ class APIErrors
 	
 	const RECORDED_NOT_READY = "RECORDED_NOT_READY;ENTRY_ID;Live entry [@ENTRY_ID@] cannot be deleted, recorded entry not in ready status yet";
 	
-	const RECORDEDING_FLOW_NOT_COMPLETE = "RECORDEDING_FLOW_NOT_COMPLETE;ENTRY_ID;Live entry [@ENTRY_ID@] cannot be deleted, recording engine still proccesing live event";
+	const RECORDING_FLOW_NOT_COMPLETE = "RECORDING_FLOW_NOT_COMPLETE;ENTRY_ID;Live entry [@ENTRY_ID@] cannot be deleted, recording engine still proccesing live event";
 	
-	const RECORDED_ENTRY_LIVE_MISMATCH = "RECORDED_ENTRY_LIVE_MISMATCH;LIVE_ENTRY_ID,RECORDE_ERNTRY_ID;Recorded entry [@RECORDE_ERNTRY_ID@] was not created while streaming to the given live entry [@LIVE_ENTRY_ID@]";
+	const RECORDED_ENTRY_LIVE_MISMATCH = "RECORDED_ENTRY_LIVE_MISMATCH;LIVE_ENTRY_ID,RECORDED_ENTRY_ID;Recorded entry [@RECORDED_ENTRY_ID@] was not created while streaming to the given live entry [@LIVE_ENTRY_ID@]";
 	
 	const CANNOT_DELETE_LIVE_ENTRY_WHILE_STREAMING = "CANNOT_DELETE_LIVE_ENTRY_WHILE_STREAMING;LIVE_ENTRY_ID;Live entry [@LIVE_ENTRY_ID@] cannot be deleted while streaming";
+	
+	const RECORDING_CONTENT_NOT_YET_SET = "RECORDING_CONTENT_NOT_YET_SET;ENTRY_ID;Entry [@ENTRY_ID@] cannot be deleted, waiting for recording engine to finish handling";
 }