+++ deployment/updates/scripts/addMediaRepurposingProfiles.php
@@ -1,23 +1,33 @@
 <?php
 
-$override = $argv[1];
-if ($override == '--help')
+require_once(__DIR__ . '/../../bootstrap.php');
+$firstArg = count($argv) > 1 ? $argv[1] : false;
+if ($firstArg == '--help')
 	return help();
 
-if ($override)
-	printInGreen("OVERRIDE mode\n");
+if ($firstArg == 'delete')
+{
+	if (count($argv) > 2) {
+		$ids = explode(',', $argv[2]);
+		foreach ($ids as $id) {
+			$profile = ScheduledTaskProfilePeer::retrieveByPK($id);
+			deleteScheduleTaskProfile($profile);
+		}
+	} else {
+		$templateProfiles = getAllTemplate();
+		printInGreen('No delete id had been given. Those are the the template profiles Ids on the admin partner:' . PHP_EOL);
+		echo implode(',',array_map(function($profile){return $profile->getId();},$templateProfiles)) . PHP_EOL;
+	}
+	return;
 
-require_once(__DIR__ . '/../../bootstrap.php');
-
-if ($override == 'renew')
-	deleteAllMRTemplate();
+}
 
 $defaultFilterValue = 1167609601; // 01/01/2007
 
-$md = MetadataProfilePeer::retrieveBySystemName('MRP', '-2');
+$md = MetadataProfilePeer::retrieveBySystemName('MRP', Partner::ADMIN_CONSOLE_PARTNER_ID);
 if (!$md)
 	$md = createMDPTemplateForMR(getXsd());
-if ($override)
+if ($firstArg == 'setNewXsd')
 	$md->setXsdData(getXsd());
 $md->save();
 
@@ -31,27 +41,25 @@ addExportAndDeleteLocalContentTemplate();
 addExportTemplate();
 
 
-printInGreen("DONE\n");
+printInGreen("DONE" . PHP_EOL);
 exit;
 
 
 
-
 function exitIfNull($val, $error) {if (!$val) exit("\n$error\n");}
-function printInGreen($str) {
-	echo "\033[32m$str\033[0m";
-}
+function printInGreen($str) {echo "\033[32m$str\033[0m";}
 
 function help() {
-	echo "This script is adding template metadata profile for Media Repurposing feature for partner -2\n";
-	echo "It also add all MR template profile for partner -2\n";
-	echo "if you want to override template of MDP run with first args as 1\n";
-	echo "if you want to delete all exist MR template and insert new ones run with first args as 'renew' \n";
+	$adminPartnerId = Partner::ADMIN_CONSOLE_PARTNER_ID;
+	echo "This script is adding template metadata profile for Media Repurposing feature for partner $adminPartnerId\n";
+	echo "It also add all MR template profile for partner $adminPartnerId\n";
+	echo "if you want to override template of MDP run with first args as 'setNewXsd'\n";
+	echo "if you want to delete exist MR template run with first argument as delete and second argument with the IDs to delete \n";
 	return 0;
 }
 
 function createMDPTemplateForMR($xsd) {
-	echo "Creating new MDP template for MR in partner -2\n";
+	echo "Creating new MDP template for MR in partner " . Partner::ADMIN_CONSOLE_PARTNER_ID . PHP_EOL;
 	$newMDP = new MetadataProfile();
 	$newMDP->setName('Media Repurposing Profiles');
 	$newMDP->setSystemName('MRP');
@@ -59,37 +67,44 @@ function createMDPTemplateForMR($xsd) {
 	$newMDP->setCreateMode(1);
 	$newMDP->setObjectType(1);
 	$newMDP->setXsdData($xsd);
-	$newMDP->setPartnerId(-2);
+	$newMDP->setPartnerId(Partner::ADMIN_CONSOLE_PARTNER_ID);
 	return $newMDP;
 }
 
-function deleteAllMRTemplate()
+function getAllTemplate()
 {
-	echo 'DELETE ALL TEMPLATE....';
 	$criteria = new Criteria();
-	$criteria->add(ScheduledTaskProfilePeer::PARTNER_ID, -2);
-	ScheduledTaskProfilePeer::doDelete($criteria);
+	$criteria->add(ScheduledTaskProfilePeer::PARTNER_ID, Partner::ADMIN_CONSOLE_PARTNER_ID);
+	return ScheduledTaskProfilePeer::doSelect($criteria);
 }
 
-function createST($name, $filter, $task)
+function deleteScheduleTaskProfile($profile)
 {
-	$st = new ScheduledTaskProfile();
-	$st->setName($name);
-	$st->setPartnerId(-2);
-	$st->setStatus(ScheduledTaskProfileStatus::DISABLED);
-	$st->setObjectFilterEngineType(ObjectFilterEngineType::ENTRY);
-	$st->setObjectFilterApiType('KalturaMediaEntryFilter');
-	$st->setObjectFilter($filter);
-	$st->setObjectTasks(array($task));
-	$st->setMaxTotalCountAllowed(500);
-	return $st;
+	/* @var $profile ScheduledTaskProfile */
+	echo 'Deleting schedule task profile [' . $profile->getId() . ']' . PHP_EOL;
+	$profile->setStatus(ScheduledTaskProfileStatus::DELETED);
+	$profile->save();
+}
+
+function createScheduleTask($name, $filter, $task)
+{
+	$scheduleTask = new ScheduledTaskProfile();
+	$scheduleTask->setName($name);
+	$scheduleTask->setPartnerId(Partner::ADMIN_CONSOLE_PARTNER_ID);
+	$scheduleTask->setStatus(ScheduledTaskProfileStatus::DISABLED);
+	$scheduleTask->setObjectFilterEngineType(ObjectFilterEngineType::ENTRY);
+	$scheduleTask->setObjectFilterApiType('KalturaMediaEntryFilter');
+	$scheduleTask->setObjectFilter($filter);
+	$scheduleTask->setObjectTasks(array($task));
+	$scheduleTask->setMaxTotalCountAllowed(500);
+	return $scheduleTask;
 }
 
 function addTask($type, $name, $filter, $description, $isFirst = false)
 {
 	$task = new kObjectTask();
 	$task->setType($type);
-	$scheduleTask = createST($name, $filter, $task);
+	$scheduleTask = createScheduleTask($name, $filter, $task);
 	$scheduleTask->setDescription($description);
 	if ($isFirst)
 		$scheduleTask->setSystemName(MediaRepurposingUtils::MEDIA_REPURPOSING_SYSTEM_NAME);
@@ -103,7 +118,7 @@ function addMailTask($name, $filter, $description, $isFirst = true)
 	$task->setType(ObjectTaskType::MAIL_NOTIFICATION);
 	$task->setDataValue('message', "You have been identified as the owner of Media which is approaching its media retention deadline. 
 		Unless you take action in the next [notification interval] days, this entry will be [Action].  ");
-	$scheduleTask = createST($name, $filter, $task);
+	$scheduleTask = createScheduleTask($name, $filter, $task);
 	$scheduleTask->setDescription($description);
 	if ($isFirst)
 		$scheduleTask->setSystemName(MediaRepurposingUtils::MEDIA_REPURPOSING_SYSTEM_NAME);
@@ -162,12 +177,14 @@ function addArchiveAndDeleteTemplate()
 	$filter = getFilter(array('_lte_created_at'));
 
 	//create the sub-task
-	$timeToNext2 = "7";
-	$task2Id = addTask(ObjectTaskType::DELETE_ENTRY, buildSubTaskName($name, "2"), $filter, $timeToNext2 );
+	$timeToNext3 = "7";
+	$task3Id = addTask(ObjectTaskType::DELETE_ENTRY, buildSubTaskName($name, "3"), $filter, $timeToNext3 );
+	$timeToNext2 = "0";
+	$task2Id = addTask(ObjectTaskType::MODIFY_ENTRY, buildSubTaskName($name, "2"), $filter, $timeToNext2 );
 	$timeToNext1 = "30";
 	$task1Id = addTask(ObjectTaskType::MODIFY_CATEGORIES, buildSubTaskName($name, "1"), $filter, $timeToNext1 );
 
-	$description = $task1Id . "[$timeToNext1]," . $task2Id . "[$timeToNext2]";
+	$description = $task1Id . "[$timeToNext1]," . $task2Id . "[$timeToNext2]," . $task3Id . "[$timeToNext3]";
 	addMailTask($name, $filter, $description);}
 
 function addArchiveExportAndDeleteTemplate()
@@ -177,14 +194,16 @@ function addArchiveExportAndDeleteTemplate()
 	$filter = getFilter(array('_lte_created_at'));
 
 	//create the sub-task
+	$timeToNext4 = "7";
+	$task4Id = addTask(ObjectTaskType::DELETE_ENTRY, buildSubTaskName($name, "3"), $filter, $timeToNext4 );
 	$timeToNext3 = "7";
-	$task3Id = addTask(ObjectTaskType::DELETE_ENTRY, buildSubTaskName($name, "3"), $filter, $timeToNext3 );
-	$timeToNext2 = "7";
-	$task2Id = addTask(ObjectTaskType::STORAGE_EXPORT, buildSubTaskName($name, "2"), $filter, $timeToNext2 );
+	$task3Id = addTask(ObjectTaskType::STORAGE_EXPORT, buildSubTaskName($name, "2"), $filter, $timeToNext3 );
+	$timeToNext2 = "0";
+	$task2Id = addTask(ObjectTaskType::MODIFY_ENTRY, buildSubTaskName($name, "2"), $filter, $timeToNext2 );
 	$timeToNext1 = "30";
 	$task1Id = addTask(ObjectTaskType::MODIFY_CATEGORIES, buildSubTaskName($name, "1"), $filter, $timeToNext1 );
 
-	$description = $task1Id . "[$timeToNext1]," . $task2Id . "[$timeToNext2]," . $task3Id . "[$timeToNext3]";
+	$description = $task1Id . "[$timeToNext1]," . $task2Id . "[$timeToNext2]," . $task3Id . "[$timeToNext3]," . $task4Id . "[$timeToNext4]";
 	addMailTask($name, $filter, $description);}
 
 function addDeletePrivateContentTemplate()