+++ plugins/content_distribution/providers/cross_kaltura/lib/model/CrossKalturaDistributionCustomDataField.php
@@ -11,4 +11,6 @@ class CrossKalturaDistributionCustomDataField
     const DISTRIBUTED_METADATA = 'CrossKalturaDistribution.distributedMetadata';
     const DISTRIBUTED_CAPTION_ASSETS = 'CrossKalturaDistribution.distributedCaptionAssets';
     const DISTRIBUTED_CUE_POINTS = 'CrossKalturaDistribution.distributedCuePoints';	
+    const DISTRIBUTED_THUMB_CUE_POINTS = 'CrossKalturaDistribution.distributedThumbCuePoints';
+    const DISTRIBUTED_TIMED_THUMB_ASSETS = 'CrossKalturaDistribution.distributedTimedThumbAssets';
 }